Create an interactive tree network plot in an htmlwidget

The tree_network function creates an interactive tree network plot in an htmlwidget

tree_network(
  data,
  width = NULL,
  height = NULL,
  treeType = "tidy",
  direction = "right",
  linkType = "diagonal",
  ...
)

Arguments

data
a tree network description in one of numerous forms (see details)

width
width of exported htmlwidget in pixels (single integer value; default == NULL)

height
height of exported htmlwidget in pixels (single integer value; default == NULL)

treeType
type of tree; one of "tidy" or "cluster" (see details) (default == "tidy")

direction
direction toward which the tree grows; one of "right", "left", "down", or "up" (see details) (default == "right")

linkType
type on link shape; one of "diagonal" or "elbow" (see details) (default == "diagonal")

...
other options (see details)

Examples

treedf <- data.frame(nodeId = LETTERS[1:7],
                     parentId = c("", "A", "A", "B", "B", "C", "C"),
                     name = LETTERS[1:7],
                     stringsAsFactors = FALSE)
tree_network(treedf)